refactor(ux): :recycle: hoist ChartTooltip to design-system
Shared by all recharts wrappers (DSA visualizers today, blog stats
charts next), so it graduates from a DSA-page-local component to a
design-system molecule under molecules/chart/.
