+++ src/components/content/data-structures-and-algorithms/recurrence-tree/recurrence-tree.tsx
@@ -10,7 +10,7 @@ import {
     XAxis,
     YAxis,
 } from "recharts";
-import { ChartTooltip } from "@/components/content/data-structures-and-algorithms/chart-tooltip";
+import { ChartTooltip } from "@/components/design-system/molecules/chart/chart-tooltip";
 
 const levels = () => {
     const n = 64;
