+++ src/components/content/data-structures-and-algorithms/space-complexity-visualizer/space-complexity-visualizer.tsx
@@ -10,7 +10,7 @@ import {
     XAxis,
     YAxis,
 } from "recharts";
-import { ChartTooltip } from "@/components/content/data-structures-and-algorithms/chart-tooltip";
+import { ChartTooltip } from "@/components/design-system/molecules/chart/chart-tooltip";
 
 const data = () => {
     const points = [];
