+++ src/components/content/data-structures-and-algorithms/stack-frame-comparison-chart/stack-frame-comparison-chart.tsx
@@ -10,7 +10,7 @@ import {
     XAxis,
     YAxis,
 } from "recharts";
-import { ChartTooltip } from "@/components/content/data-structures-and-algorithms/chart-tooltip";
+import { ChartTooltip } from "@/components/design-system/molecules/chart/chart-tooltip";
 
 type StackFrameData = {
     step: number;
